fix #738
