+++ src/components/Main/Navigation/NavigationContent/UsersElement.vue
@@ -16,8 +16,8 @@ import { User } from '@traptitech/traq'
 import { makeStyles } from '@/lib/styles'
 import UserIcon from '@/components/UI/UserIcon.vue'
 import UsersElementUserName from './UsersElementUserName.vue'
-import useChannelSelect from '@/use/channelSelect'
 import store from '@/store'
+import { changeDMChannelByUsername } from '@/router/channel'
 
 export default defineComponent({
   name: 'UsersElement',
@@ -45,13 +45,11 @@ export default defineComponent({
         : false
     )
 
-    const { onDMChannelSelect } = useChannelSelect()
-
     const onClick = () => {
       if (props.user.bot && props.user.name.startsWith('Webhook#')) {
         return
       }
-      onDMChannelSelect(props.user.name, props.user.id)
+      changeDMChannelByUsername(props.user.name)
     }
 
     return {
