+++ src/views/use/routeWatcher.ts
@@ -5,8 +5,26 @@ import useNavigationController from '@/use/navigationController'
 import useChannelPath from '@/use/channelPath'
 import useViewTitle from './viewTitle'
 import apis from '@/lib/apis'
+import { ChannelId, DMChannelId } from '@/types/entity-ids'
 
 type Views = 'none' | 'main' | 'not-found'
+
+const setUnreadState = (id: ChannelId | DMChannelId) => {
+  // 未読の処理
+  // TODO: 新着メッセージ基準設定などの処理
+  store.commit.domain.messagesView.unsetUnreadSince()
+  const unreadChannel = store.state.domain.me.unreadChannelsSet[id]
+  if (unreadChannel) {
+    if (
+      store.state.domain.me.subscriptionMap[id] > 0 ||
+      store.state.entities.dmChannels[id]
+    ) {
+      store.commit.domain.messagesView.setUnreadSince(unreadChannel.since)
+    }
+
+    store.dispatch.domain.me.readChannel({ channelId: id })
+  }
+}
 
 const useRouteWacher = (context: SetupContext) => {
   const { channelPathToId, channelIdToPathString } = useChannelPath()
@@ -66,6 +84,9 @@ const useRouteWacher = (context: SetupContext) => {
         state.channelParam.split('/'),
         store.state.domain.channelTree.channelTree
       )
+
+      setUnreadState(id)
+
       store.dispatch.ui.mainView.changePrimaryViewToChannel({
         channelId: id,
         entryMessageId: context.root.$route.query?.message as string
@@ -82,13 +103,23 @@ const useRouteWacher = (context: SetupContext) => {
     const user = store.getters.entities.userByName(state.currentRouteParam)
     try {
       if (!user) throw 'user not found'
-      const res = await apis.getUserDMChannel(user.id)
-      store.commit.entities.addDMChannel({
-        id: res.data.id,
-        entity: res.data
-      })
+
+      let dmChannelId = store.getters.entities.DMChannelIdByUserId(user.id)
+
+      if (!dmChannelId) {
+        const { data } = await apis.getUserDMChannel(user.id)
+        store.commit.entities.addDMChannel({
+          id: data.id,
+          entity: data
+        })
+        dmChannelId = data.id
+      }
+      if (!dmChannelId) throw 'failed to fetch DM channel ID'
+
+      setUnreadState(dmChannelId)
+
       store.dispatch.ui.mainView.changePrimaryViewToDM({
-        channelId: res.data.id,
+        channelId: dmChannelId,
         userName: user.name,
         entryMessageId: context.root.$route.query?.message as string
       })
